requirements.txt
flask==3.0.0
google-genai==0.1.0
openai==1.12.0
PyPDF2==3.0.1
